requirements.txt
pandas
google-cloud-storage
flask
gunicorn
gcsfs
binance-sdk-spot